chore: moved helper file outside src
