+++ src/routes/users/index.ts
@@ -2,7 +2,7 @@ import express from 'express'
 const router = express.Router()
 const { getAllUsers, addUser, updateUser, deleteUser } = require('./default')
 import { auth } from './customPost'
-import { isUserLoggedIn } from '../../helper'
+import { isUserLoggedIn } from '../../../helper'
 
 //default
 router.get('/', isUserLoggedIn, getAllUsers)
